[refactor] move the get_highest_match function into license.py instead of being duplicated in both github and gitlab.
